+++ app/src/main/java/org/eyeseetea/malariacare/MalariaCareDbHelper.java
@@ -1,4 +1,4 @@
-package org.psi.malariacare;
+package org.eyeseetea.malariacare;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
