+++ app/src/main/java/org/eyeseetea/malariacare/questions/Questions.java
@@ -1,4 +1,4 @@
-package org.psi.malariacare.dummy;
+package org.eyeseetea.malariacare.questions;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -11,26 +11,26 @@ import java.util.Map;
  * <p/>
  * TODO: Replace all uses of this class before publishing your app.
  */
-public class DummyContent {
+public class Questions {
 
     /**
      * An array of sample (dummy) items.
      */
-    public static List<DummyItem> ITEMS = new ArrayList<DummyItem>();
+    public static List<Question> ITEMS = new ArrayList<>();
 
     /**
      * A map of sample (dummy) items, by ID.
      */
-    public static Map<String, DummyItem> ITEM_MAP = new HashMap<String, DummyItem>();
+    public static Map<String, Question> ITEM_MAP = new HashMap<>();
 
     static {
         // Add 3 sample items.
-        addItem(new DummyItem("1", "Item 1"));
-        addItem(new DummyItem("2", "Item 2"));
-        addItem(new DummyItem("3", "Item 3"));
+        addItem(new Question("1", "Item 1"));
+        addItem(new Question("2", "Item 2"));
+        addItem(new Question("3", "Item 3"));
     }
 
-    private static void addItem(DummyItem item) {
+    private static void addItem(Question item) {
         ITEMS.add(item);
         ITEM_MAP.put(item.id, item);
     }
@@ -38,11 +38,11 @@ public class DummyContent {
     /**
      * A dummy item representing a piece of content.
      */
-    public static class DummyItem {
+    public static class Question {
         public String id;
         public String content;
 
-        public DummyItem(String id, String content) {
+        public Question(String id, String content) {
             this.id = id;
             this.content = content;
         }
